requirements.txt
# Core Dependencies
azure-devops-python-api>=6.0.0
requests>=2.28.2
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.3.1
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.3.0
mypy>=1.3.0

# Optional: Logging and Debugging
logging>=0.5.1.2

# Testing
mock>=5.0.2

# Documentation
sphinx>=6.2.1
sphinx-rtd-theme>=1.2.0

# Type Checking
types-requests>=2.28.11
